Fashion_Item_Classification_using_Convolutional Neural Networks/requirements.txt
fastapi
uvicorn[standard]
tensorflow
keras
numpy
pillow
aiofiles
scikit-image
